allow sections to be added from sidebar
